requirements.txt
Flask
Flask-MySQLdb
mysql-connector-python
passlib
WTForms
email_validator
